+++ src/loaders/core-modules.ts
@@ -15,3 +15,4 @@ export { applyViewVisibility } from '../utils/view-visibility';
 export { insertCustomViews, resolveCustomViews } from '../utils/custom-view-ref';
 export { applyDesign } from '../utils/design';
 export { isUtilityViewEnabled } from '../utils/summary-view-utils';
+export { resolveAutomaticFeatures } from '../utils/feature-availability';